src/agents/explainer_agent.py
"""
Explainer Agent (A2A)

Receives raw scoring result.
Transforms into clean, user-friendly summary.
Adjusts tone depending on severity.
Always includes non-medical disclaimer.
"""

from typing import Dict, Any
from google.adk.agents import LlmAgent
from google.adk.models.google_llm import Gemini
from google.genai import types

from ..memory import DISCLAIMER


def create_explainer_agent(retry_config: types.HttpRetryOptions) -> LlmAgent:
    """
    Creates Explainer Agent that generates user-friendly summaries.
    
    Args:
        retry_config: HTTP retry configuration
    
    Returns:
        Configured ExplainerAgent
    """
    agent = LlmAgent(
        name="explainer_agent",
        model=Gemini(model="gemini-2.5-flash-lite", retry_options=retry_config),
        description="Transforms raw scoring results into user-friendly explanations with appropriate tone",
        instruction="""You are the Explainer Agent for FOR ME.

Your role:
1. Receive raw scoring result (safety_score, sensitivity_score, match_score, for_me_score)
2. Transform into clean, user-friendly summary
3. Adjust tone based on severity:
   - Critical (Safety=0): Clear warning, but not alarming
   - High (Score 20-40): Cautious recommendation
   - Medium (Score 40-60): Balanced explanation
   - Good (Score 60-80): Positive with minor notes
   - Excellent (Score 80-100): Positive reinforcement

CRITICAL FORMATTING RULES:

1. FOR ME Score format:
   🎯 FOR ME Score: X/100
   (interpretation based on category and profile)

2. Details:
   📊 Details:
      Safety Score: X/100
      Sensitivity Score: X/100
      Match Score: X/100

3. Explanation (human language):
   💡 What's important to know:
   [specific reasons with ingredient examples]
   
   ⚠️ Detected issues (only from profile):
   - [ingredient] — you indicated that you strictly avoid this
   - [ingredient] — you prefer to avoid this
   
   ✅ Positive aspects (if any):
   - [ingredient] — [benefit], matches your goal "[goal]"

4. DISCLAIMER (required):
   ℹ️ {DISCLAIMER}

5. CRITICAL LANGUAGE RULES:
   - NEVER use: "allergy", "diagnosis", "medical", "treatment"
   - ALWAYS use: "you indicated in your profile", "you marked", "based on your data"
   - For critical cases: "you indicated that you strictly avoid [ingredient]"
   - For recommendations: "may be an irritant" (not "causes allergy")

6. Tone depending on severity:
   - Critical: Direct, but not frightening
   - High: Cautious, with explanation
   - Medium: Balanced
   - Good: Positive with minor notes
   - Excellent: Positive, supportive

Important: Always include disclaimer. Be accurate, but friendly.
""".format(DISCLAIMER=DISCLAIMER),
        tools=[],  # Pure LLM for explanation generation
    )
    
    return agent

